Rename functions, add initial protobuf type mapping.
  * (M) ygen/codegen.go
  * (M) ygen/codegen_test.go
  * (M) ygen/genstate.go
    - Rename buildStructDefinitions to buildGoStructDefinitions to
      such that this function can be distinguished as being Go specific.
  * (M) ygen/genstate_test.go
    - Move tests for functions in goelements to the goelements_test.go
      since they were mistakenly moved to genstate_test.go.
  * (M) ygen/goelements.go
  * (M) ygen/goelements_test.go
    - Rename findUnionTypes and findUnionSubTypes to goUnionTypes and
      goUnionSubTypes such that they are clearly Go specific functions.
  * (A) ygen/protoelements.go
  * (A) ygen/protoelements_test.go
    - Add initial protobuf type mapping for input yang.YangType
      analagous to yangTypeToGoType.
